statement: rename query.rs -> unprepared.rs
