+++ docs/source/tracing/basic.md
@@ -10,7 +10,7 @@ return a `QueryResult` which contains a `tracing_id` if tracing was enabled.
 # use scylla::client::session::Session;
 # use std::error::Error;
 # async fn check_only_compiles(session: &Session) -> Result<(), Box<dyn Error>> {
-use scylla::statement::query::Statement;
+use scylla::statement::unprepared::Statement;
 use scylla::response::query_result::QueryResult;
 use scylla::observability::tracing::TracingInfo;
 use uuid::Uuid;
