+++ docs/source/tracing/prepare.md
@@ -8,7 +8,7 @@
 # use scylla::client::session::Session;
 # use std::error::Error;
 # async fn check_only_compiles(session: &Session) -> Result<(), Box<dyn Error>> {
-use scylla::statement::query::Statement;
+use scylla::statement::unprepared::Statement;
 use scylla::statement::prepared::PreparedStatement;
 use scylla::observability::tracing::TracingInfo;
 use uuid::Uuid;
