+++ scylla/src/statement/mod.rs
@@ -15,9 +15,10 @@ use crate::policies::retry::RetryPolicy;
 
 pub mod batch;
 pub mod prepared;
-pub mod query;
+pub mod unprepared;
 
 pub use crate::frame::types::{Consistency, SerialConsistency};
+pub use unprepared::Statement;
 
 // This is the default common to drivers.
 const DEFAULT_PAGE_SIZE: i32 = 5000;
